Significant refactor that removes the replisome plugin-specific code from the main receiver/filter/consumer pipeline flow. Extracted generic DB configure and access code from original JsonReceiver class. Created new TestDecodingReceiver that works with basic test_decoding plugin available in default Postgresql install. Minor change to support internal receiver/filter/consumer class names without assuming specific module names.
