+++ setup.py
@@ -1,3 +1,4 @@
+#!/usr/bin/env python
 import os
 import re
 import sys
